Structure updated
