+++ pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
     "polyfactory==2.19.0",
     "pydantic==2.10.6",
     "pydantic-core==2.27.2",
+    "pydantic-settings>=2.8.1",
     "pygments==2.19.1",
     "python-dotenv>=1.0.1",
     "pyyaml==6.0.2",
