+++ src/application/services/payment_service.py
@@ -1,0 +1,1 @@
+class PaymentService: ...
