--- src/infrastructure/database/base.py
@@ -1,21 +1,0 @@
-from uuid import UUID
-
-from sqlalchemy import MetaData
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
-
-
-class _BaseModel(DeclarativeBase):
-    metadata = MetaData(
-        naming_convention={
-            "ix": "ix_%(column_0_label)s",
-            "uq": "uq_%(table_name)s_%(column_0_name)s",
-            "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-            "pk": "pk_%(table_name)s",
-        }
-    )
-
-
-class Base(_BaseModel):
-    __abstract__ = True
-
-    id: Mapped[UUID] = mapped_column(primary_key=True)
