+++ src/infrastructure/database/config.py
@@ -1,19 +1,37 @@
+from uuid import UUID
+
 from litestar.contrib.sqlalchemy.plugins import (
     SQLAlchemyAsyncConfig,
     SQLAlchemyPlugin,
 )
+from sqlalchemy import MetaData
+from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
-from src.infrastructure.database.base import Base
 from src.infrastructure.database.settings import settings
+
+
+class _BaseModel(DeclarativeBase):
+    metadata = MetaData(
+        naming_convention={
+            "ix": "ix_%(column_0_label)s",
+            "uq": "uq_%(table_name)s_%(column_0_name)s",
+            "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
+            "pk": "pk_%(table_name)s",
+        }
+    )
+
+
+class Base(_BaseModel):
+    __abstract__ = True
+    id: Mapped[UUID] = mapped_column(primary_key=True)
 
 
 def get_sqlalchemy_config() -> SQLAlchemyAsyncConfig:
     """Get SQLAlchemy config."""
     return SQLAlchemyAsyncConfig(
-        connection_string=settings.database_url,
-        create_all=True,
-        metadata=Base.metadata
+        connection_string=settings.database_url, create_all=True, metadata=Base.metadata
     )
+
 
 def get_sqlalchemy_plugin() -> SQLAlchemyPlugin:
     """Get SQLAlchemy plugin."""
