+++ src/interfaces/api/controllers/controllers.py
@@ -1,0 +1,7 @@
+from litestar import Controller
+
+
+class PaymentController(Controller):
+    path = "/payments"
+
+    async def create_payment(self) -> None: ...
